Use utils as static crawler and save downloaded info into a file
